pyproject.toml
[project]
name = "pk-timetable"
version = "0.1.0"
description = "Sync university timetable from .xlsx to Google Calendar"
requires-python = ">=3.11"
dependencies = [
    "openpyxl>=3.1",
    "requests>=2.32",
    "google-api-python-client>=2.150",
    "google-auth>=2.37",
    "pydantic>=2.10",
    "pyyaml>=6.0",
]

[project.scripts]
pk-timetable = "pk_timetable.main:main"

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pk_timetable"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "hatchling>=1.29.0",
    "pytest>=8.0",
    "pytest-mock>=3.14",
]
